% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{addDataSetFromHub}
\alias{addDataSetFromHub}
\title{Add dataSet from AnnotationHub}
\usage{
addDataSetFromHub(OGREDataSet, dataSet, type)
}
\arguments{
\item{OGREDataSet}{OGREDataSet}

\item{dataSet}{\code{character} Name of one predefined dataSets to add as
query or subject to a OGREDataSet. Possible dataSets can be show with
\code{listPredefinedDataSets()}.}

\item{type}{Type of dataSet, must be either query or subject. If query the
dataSet will be added as query and at the first position of OGREDataSet.}
}
\value{
OGREDataSet.
}
\description{
AnnotationHub offers a wide range of annotated datasets which can be manually
aquired but need some parsing to work with OGRE as detailed in vignette
section "Load datasets from AnnotationHub".
For convienence \code{addDataSetFromHub()} adds one of the predefined human
dataSets of \code{listPredefinedDataSets()} to a OGREDataSet.Those are taken from
AnnotationHub and are ready to use for OGRE. Additional information on
dataSets can be found here \code{\link{listPredefinedDataSets}}.
}
\examples{
myOGRE <- OGREDataSet() 
myOGRE <- addDataSetFromHub(myOGRE,"protCodingGenes","query")
}
