% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{listPredefinedDataSets}
\alias{listPredefinedDataSets}
\title{List predefined datasets}
\usage{
listPredefinedDataSets()
}
\value{
\code{character} vector.
}
\description{
Use \code{listPredifinedDataSets()} to receive a vector of names for predefined
datasets that can be aquired from AnnotationHub that are already correctly
parsed and formatted. Each of the listed names can be used as input for
\code{addDataSetFromHub()}.
}
\examples{
listPredefinedDataSets()
}
