% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{loadAnnotations}
\alias{loadAnnotations}
\title{loadAnnotations}
\usage{
loadAnnotations(TREGELDataSet)
}
\arguments{
\item{TREGELDataSet}{A TREGELDataSet.}
}
\value{
TREGELDataSet.
}
\description{
loadAnnotations
}
