% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{makeExampleGRanges}
\alias{makeExampleGRanges}
\title{Make an example GRanges dataset}
\usage{
makeExampleGRanges()
}
\value{
OGREDataSet.
}
\description{
\code{makeExampleGRanges} generates an example GRanges dataset.
}
\examples{
myGRanges <- makeExampleGRanges()
}
