% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{makeExampleOGREDataSet}
\alias{makeExampleOGREDataSet}
\title{Make a example OGRE dataset}
\usage{
makeExampleOGREDataSet()
}
\value{
OGREDataSet.
}
\description{
\code{makeExampleOGREDataSet} generates a example OGREDataSet from dataset files
stored in OGRE's extdata directory.
}
\examples{
myOGRE=makeExampleOGREDataSet()
}
