% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{OGREDataSet}
\alias{OGREDataSet}
\title{BuildOGREDataSet}
\usage{
OGREDataSet()
}
\value{
A OGREDataSet.
}
\description{
Builds a \code{OGREDataset} as a \code{GenomicRangesList} for storing and analysing
datasets which can be added by \code{addDataSetFromHub()} or \code{addGRanges()}.
Use \code{BuildOGREDataSetFromDir} for adding dataSets stored as files.
}
\examples{
myOGRE <- OGREDataSet()
}
