% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{readDataSetFromFolder}
\alias{readDataSetFromFolder}
\title{Read dataset(s) from folder}
\usage{
readDataSetFromFolder(OGREDataSet, type)
}
\arguments{
\item{OGREDataSet}{A OGREDataSet.}

\item{type}{\code{character} and one of query/subject.}
}
\value{
A OGREDataSet.
}
\description{
\code{\link[=readDataSetFromFolder]{readDataSetFromFolder()}} scanns queryFolder and subjectFolder for either
.RDS/.rds or .CSV/.csv files and adds them to a OGREDataSet. Each region
needs chromosome, start, end and strand information. (tabular file columns
must be named accordingly). A unique ID and a name column must be present in
the \code{GenomicRanges} object's metatdata and tabular file.
}
\examples{
myOGRE <- makeExampleOGREDataSet()
myOGRE <- readDataSetFromFolder(myOGRE,type="query")
myOGRE <- readDataSetFromFolder(myOGRE,type="subject")
}
