% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{subsetGRanges}
\alias{subsetGRanges}
\title{Subset a GRanges object}
\usage{
subsetGRanges(OGREDataSet, IDs, name)
}
\arguments{
\item{OGREDataSet}{An OGREDataSet}

\item{IDs}{\code{character vector} with IDs used to subset the GRanges object
defined in \code{name}}

\item{name}{\code{character} Name of the GRanges object for subsetting. One of
the GRanges objects in a \code{OGREDataSet}}
}
\value{
OGREDataSet.
}
\description{
Subsets a GRanges object with reference to it's ID column using a ID vector.
}
\examples{
myOGRE <- makeExampleOGREDataSet()
myOGRE <- loadAnnotations(myOGRE)
myOGRE <- subsetGRanges(myOGRE,c("ENSG00000142168","ENSG00000256715"),"genes")
}
