% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{sumPlot}
\alias{sumPlot}
\title{Generate summary plot}
\usage{
sumPlot(TREGELDataSet)
}
\arguments{
\item{TREGELDataSet}{A TREGELDataSet.}
}
\value{
TREGELDataSet.
}
\description{
\code{sumPlot()} calculates key numbers i.e. (total number of overlaps, number of overlaps per subject...) to help with an
exploratory data evaluation and displays them in an informative barplot.
}
