% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{addDataSetFromHub}
\alias{addDataSetFromHub}
\title{Add dataset from AnnotationHub}
\usage{
addDataSetFromHub(OGREDataSet, dataSet, type)
}
\arguments{
\item{OGREDataSet}{OGREDataSet}

\item{dataSet}{\code{character} Name of one predefined dataSets to add as
query or subject to a OGREDataSet. Possible dataSets can be show with
\code{listPredefinedDataSets()}}

\item{type}{Type of dataSet, must be either query or subject. If query the
dataSet will be added as query and at the first position of OGREDataSet.}
}
\value{
OGREDataSet.
}
\description{
Adds one of the predifined human dataSets of \code{listPredefinedDataSets()} to a
OGREDataSet. DataSets are taken from AnnotationHub and are ready to use for
OGRE.
}
\examples{
myOGRE <- OGREDataSet() 
myOGRE <- addDataSetFromHub(myOGRE,"protCodingGenes","query")
}
