% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{addGRanges}
\alias{addGRanges}
\title{Add a GenomicRanges dataset to OGREDataSet}
\usage{
addGRanges(OGREDataSet, dataSet, type)
}
\arguments{
\item{dataSet}{A GRanges object Each region needs chromosome, start, end and
strand information. A unique ID and a name column must be present in the
\code{GenomicRanges} object metadata.}

\item{type}{Type of dataSet, must be either query or subject. If query the
dataSet will be added as query and at the first position of OGREDataSet.}
}
\value{
OGREDataSet.
}
\description{
Add a GenomicRanges dataset to OGREDataSet
}
\examples{
myOGRE <- OGREDataSet()
gr <- GRanges(Rle(c("chr2", "chr2", "chr1", "chr3"), c(1, 3, 2, 4)),
      IRanges(1:10, width=10:1, names=head(letters, 10)),
      Rle(strand(c("-", "+", "*", "+", "-")), c(1, 2, 2, 3, 2)),
      ID=1:10, name=paste0("gene",1:10))
myOGRE <- addGRanges(myOGRE,gr,"query")
}
