% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fOverlaps}
\alias{fOverlaps}
\title{Finds all overlaps between query and subject(s) and stores hits in two separate files ''}
\usage{
fOverlaps(TREGELDataSet)
}
\arguments{
\item{TREGELDataSet}{A TREGELDataSet.}
}
\value{
TREGELDataSet.
}
\description{
Finds all overlaps between query and subject(s) and stores hits in two separate files ''
}
