% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fOverlaps}
\alias{fOverlaps}
\title{Find overlaps}
\usage{
fOverlaps(OGREDataSet, selfHits = FALSE)
}
\arguments{
\item{OGREDataSet}{A OGREDataSet.}

\item{selfHits}{\code{logical} if FALSE(default) ignores self hits of
identical regions within datasets.}
}
\value{
OGREDataSet.
}
\description{
Finds all overlaps between query and subject(s) and stores each hit (overlap)
in data table \code{detailDT}. Data table \code{sumDT} shows all
overlaps of a certain subject type for all query elements. By default also
partially overlaps are reported.
}
\examples{
myOGRE <- makeExampleOGREDataSet()
myOGRE <- loadAnnotations(myOGRE)
myOGRE <- fOverlaps(myOGRE)
}
