% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{gvizPlot}
\alias{gvizPlot}
\title{Gviz plotting}
\usage{
gvizPlot(
  TREGELDataSet,
  query,
  trackRegionLabels = setNames(rep("ID", length(TREGELDataSet)), names(TREGELDataSet))
)
}
\description{
Gviz plotting
}
