% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{readSubject}
\alias{readSubject}
\title{Read subject datasets}
\usage{
readSubject(TREGELDataSet)
}
\arguments{
\item{TREGELDataSet}{A TREGELDataSet.}
}
\value{
A TREGELDataSet.
}
\description{
\code{\link[=readSubject]{readSubject()}} scanns \code{SubjectFolder} for any \code{GRanges} objects stored as .RDS / .rds files and attaches them to the TREGELDataSet.
}
